backend v1

+++ api/Controllers/CommentController.cs
@@ -5,8 +5,6 @@ using System.Threading.Tasks;
 using api.Dtos.Comment;
 using api.Interfaces;
 using api.Mappers;
-using api.Models;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 
 namespace api.Controllers
@@ -26,6 +24,9 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -33,44 +34,64 @@ namespace api.Controllers
             return Ok(commentDto);
         }
 
-        [HttpPost("{stockId}")]
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
+        {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var comment = await _commentRepo.GetByIdAsync(id);
+
+            if (comment == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(comment.ToCommentDto());
+        }
+
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
-            if(!await _stockRepo.StockExists(stockId))
+            if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
             }
+
             var commentModel = commentDto.ToCommentFromCreate(stockId);
             await _commentRepo.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById), new { id = commentDto.Id}, commentModel.ToCommentDto());
+            return CreatedAtAction(nameof(GetById), new { id = commentModel }, commentModel.ToCommentDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto);
+        [Route("{id:int}")] 
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
-            var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdateDto());
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
 
-            if(Comment == null)
+            if (comment == null)
             {
                 return NotFound("Comment not found");
             }
 
-            return Ok(Comment.ToCommentDto());
+            return Ok(comment.ToCommentDto());
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var commentModel1 = await _commentRepo.DeleteAsync(id);
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var commentModel = await _commentRepo.DeleteAsync(id);
 
-            if(commentModel1 == null)
+            if (commentModel == null)
             {
                 return NotFound("Comment does not exist");
             }
 
-            return Ok(commentModel1);
+            return Ok(commentModel);
         }
     }
 }
